Refactor main function (#5980)
* node: Refactor main execution flow and introduce launcher module

* node/launcher: extract setup_configuration helper  from run

* node/launcher: extract setup_metrics helper  from run

* node/launcher: extract setup_store helper  from run

* node/launcher: extract build_blockchain_map helper  from run

* node/launcher: extract cleanup_ethereum_shallow_blocks helper  from run

* node/launcher: extract spawn_block_ingestor helper  from run

* node/launcher: extract deploy_subgraph_from_flag helper  from run

* node/launcher: extract spawn_contention_checker helper  from run

* node/launcher: extract build_graphql_server helper  from run

* node/launcher: extract build_subgraph_registrar helper  from run
